+++ Models/Entities/Tbl_Factor.cs
@@ -10,6 +10,6 @@ namespace sabashop.Models.Entities
         public long Price { get; set; }
         public bool StatusPay { get; set; }
         public DateTime PayTime { get; set; }
-        public int RFactor { get; set; }
+        public long RFactor { get; set; }
     }
 }
